+++ datasetCreation/scripts/DP10_DataAugmentation.py
@@ -1,3 +1,4 @@
+import numpy as np
 from numpy import sqrt, asarray
 
 def chunker(seq, size):
@@ -34,7 +35,7 @@ def DP10_getDistanceVector2Poses(p1,p2):
         dist = sqrt( (x1-x2)**2 + (y1-y2)**2 )
         out.append(dist)
 
-    out = asarray(out, dtype=float, order=None)
+    out = asarray(out, dtype=np.float32, order=None)
     return out
    
 def DP10_getMotionVector2Poses(p1,p2):
@@ -65,7 +66,7 @@ def DP10_getMotionVector2Poses(p1,p2):
         dist = sqrt( (x1-x2)**2 + (y1-y2)**2 )
         out.append(dist)
 
-    out = asarray(out, dtype=float, order=None)
+    out = asarray(out, dtype=np.float32, order=None)
     return out
 
 if __name__ == '__main__':
